+++ groups/tests/test_managers.py
@@ -1,8 +1,12 @@
+import datetime
+
+from django.contrib.auth import get_user_model
+from django.db import models as django_models
 from django.test import TestCase
 from incuna_test_utils.compat import Python2AssertMixin
 
 from . import factories
-from .. import models
+from .. import managers, models
 
 
 class TestGroupManager(Python2AssertMixin, TestCase):
@@ -34,6 +38,48 @@ class TestGroupManager(Python2AssertMixin, TestCase):
         factories.TextCommentFactory.create_batch(2, user=user)
 
         self.assertCountEqual([user], models.Group.objects.users())
+
+    def test_within_days(self):
+        comment = factories.TextCommentFactory.create(date_created=datetime.date.today())
+        factories.TextCommentFactory.create(date_created=datetime.date(1970, 1, 1))
+
+        results = models.Group.objects.within_days()
+        self.assertCountEqual([comment.discussion.group], results)
+
+    def test_within_time(self):
+        comment = factories.TextCommentFactory.create()
+        factories.TextCommentFactory.create(date_created=datetime.date(1970, 1, 1))
+        delta = datetime.timedelta(hours=6)
+
+        results = models.Group.objects.within_time(delta)
+        self.assertCountEqual([comment.discussion.group], results)
+
+    def test_since(self):
+        comment = factories.TextCommentFactory.create()
+        factories.TextCommentFactory.create(date_created=datetime.date(1970, 1, 1))
+        when = datetime.date(2000, 1, 1)
+
+        results = models.Group.objects.since(when)
+        self.assertCountEqual([comment.discussion.group], results)
+
+    def test_within_days_distinct(self):
+        """Assert that Group.objects.within_days() contains no duplicates."""
+        group = factories.GroupFactory.create()
+        factories.TextCommentFactory.create_batch(
+            2,
+            date_created=datetime.date.today(),
+            discussion__group=group,
+        )
+
+        self.assertCountEqual([group], models.Group.objects.within_days())
+
+    def test_within_time_distinct(self):
+        """Assert that Group.objects.within_time() contains no duplicates."""
+        group = factories.GroupFactory.create()
+        factories.TextCommentFactory.create_batch(2, discussion__group=group)
+        delta = datetime.timedelta(hours=6)
+
+        self.assertCountEqual([group], models.Group.objects.within_time(delta))
 
 
 class TestDiscussionManager(Python2AssertMixin, TestCase):
@@ -76,6 +122,48 @@ class TestDiscussionManager(Python2AssertMixin, TestCase):
         results = models.Discussion.objects.for_group(group).comments()
         self.assertCountEqual([comment], results)
 
+    def test_within_days(self):
+        comment = factories.TextCommentFactory.create(date_created=datetime.date.today())
+        factories.TextCommentFactory.create(date_created=datetime.date(1970, 1, 1))
+
+        results = models.Discussion.objects.within_days()
+        self.assertCountEqual([comment.discussion], results)
+
+    def test_within_time(self):
+        comment = factories.TextCommentFactory.create()
+        factories.TextCommentFactory.create(date_created=datetime.date(1970, 1, 1))
+        delta = datetime.timedelta(hours=6)
+
+        results = models.Discussion.objects.within_time(delta)
+        self.assertCountEqual([comment.discussion], results)
+
+    def test_since(self):
+        comment = factories.TextCommentFactory.create()
+        factories.TextCommentFactory.create(date_created=datetime.date(1970, 1, 1))
+        when = datetime.date(2000, 1, 1)
+
+        results = models.Discussion.objects.since(when)
+        self.assertCountEqual([comment.discussion], results)
+
+    def test_within_days_distinct(self):
+        """Assert that Discussion.objects.within_days() contains no duplicates."""
+        discussion = factories.DiscussionFactory.create()
+        factories.TextCommentFactory.create_batch(
+            2,
+            date_created=datetime.date.today(),
+            discussion=discussion,
+        )
+
+        self.assertCountEqual([discussion], models.Discussion.objects.within_days())
+
+    def test_within_time_distinct(self):
+        """Assert that Discussion.objects.within_time() contains no duplicates."""
+        discussion = factories.DiscussionFactory.create()
+        factories.TextCommentFactory.create_batch(2, discussion=discussion)
+        delta = datetime.timedelta(hours=6)
+
+        self.assertCountEqual([discussion], models.Discussion.objects.within_time(delta))
+
 
 class TestCommentManager(Python2AssertMixin, TestCase):
     def test_for_group(self):
@@ -107,6 +195,27 @@ class TestCommentManager(Python2AssertMixin, TestCase):
 
         self.assertCountEqual([user], models.BaseComment.objects.users())
 
+    def test_within_days(self):
+        comment = factories.TextCommentFactory.create(date_created=datetime.date.today())
+        factories.TextCommentFactory.create(date_created=datetime.date(1970, 1, 1))
+
+        self.assertCountEqual([comment], models.BaseComment.objects.within_days())
+
+    def test_within_time(self):
+        comment = factories.TextCommentFactory.create()
+        factories.TextCommentFactory.create(date_created=datetime.date(1970, 1, 1))
+        delta = datetime.timedelta(hours=6)
+
+        self.assertCountEqual([comment], models.BaseComment.objects.within_time(delta))
+
+    def test_since(self):
+        comment = factories.TextCommentFactory.create()
+        factories.TextCommentFactory.create(date_created=datetime.date(1970, 1, 1))
+        when = datetime.date(2000, 1, 1)
+
+        results = models.BaseComment.objects.since(when)
+        self.assertCountEqual([comment], results)
+
     def test_chaining(self):
         """Assert that chaining some of the above methods together works properly."""
         comment = factories.TextCommentFactory.create()
@@ -125,3 +234,74 @@ class TestCommentManager(Python2AssertMixin, TestCase):
 
         may_delete_values = [comment.user_may_delete for comment in results]
         self.assertCountEqual([True, False], may_delete_values)
+
+
+class TestWithinDaysQuerySetMixin(Python2AssertMixin, TestCase):
+    mixin = managers.WithinDaysQuerySetMixin
+
+    def __init__(self, *args, **kwargs):
+        """
+        Declare some classes that will be used in this test.
+
+        We need a manager and a queryset that inherit from the mixin being tested,
+        and a proxy model that declares the mixed-in manager as its `objects` attribute
+        so that we can call the manager methods via ProxyModel.objects.method().
+
+        This mixin is intended for use with User models, so we'll proxy that.
+        """
+        super(TestWithinDaysQuerySetMixin, self).__init__(*args, **kwargs)
+
+        class MixedInQuerySet(self.mixin, django_models.QuerySet):
+            """A basic QuerySet extending the mixin."""
+            since_filter = 'comments__date_created__gte'
+
+        class MixedInManager(self.mixin, django_models.Manager):
+            """A basic Manager extending the mixin and using MixedInQuerySet."""
+            since_filter = 'comments__date_created__gte'
+
+            def get_queryset(self):
+                return MixedInQuerySet(self.model, using=self._db)
+
+        class ProxyUser(get_user_model()):
+            """A proxy for the User model that adds in our manager."""
+            objects = MixedInManager()
+
+            class Meta:
+                proxy = True
+
+        self.model = ProxyUser
+
+    def setUp(self):
+        self.recent_user = factories.UserFactory.create()
+        factories.TextCommentFactory.create(user=self.recent_user)
+        factories.TextCommentFactory.create(date_created=datetime.date(1970, 1, 1))
+
+    def test_within_days_in_manager(self):
+        """Assert that within_days() works properly when called from the manager."""
+        results = self.model.objects.within_days()
+        self.assertCountEqual([self.recent_user], results)
+
+    def test_within_days_in_queryset(self):
+        """Assert that within_days() works properly when called from the queryset."""
+        results = self.model.objects.all().within_days()
+        self.assertCountEqual([self.recent_user], results)
+
+    def test_within_time_in_manager(self):
+        """Assert that within_time() works properly when called from the manager."""
+        results = self.model.objects.within_time(datetime.timedelta(hours=6))
+        self.assertCountEqual([self.recent_user], results)
+
+    def test_within_time_in_queryset(self):
+        """Assert that within_time() works properly when called from the queryset."""
+        results = self.model.objects.all().within_time(datetime.timedelta(hours=6))
+        self.assertCountEqual([self.recent_user], results)
+
+    def test_since_in_manager(self):
+        """Assert that since() works properly when called from the manager."""
+        results = self.model.objects.since(datetime.date(2000, 1, 1))
+        self.assertCountEqual([self.recent_user], results)
+
+    def test_since_in_queryset(self):
+        """Assert that since() works properly when called from the queryset."""
+        results = self.model.objects.all().since(datetime.date(2000, 1, 1))
+        self.assertCountEqual([self.recent_user], results)